refactor: rename fim.ts to FillInTheMiddle.ts
- Rename file for better clarity and consistency
- Update import in GhostInlineCompletionProvider
- All tests pass (63/63)
