+++ src/services/ghost/classic-auto-complete/FillInTheMiddle.ts
